requirements.txt
aerich==0.2.3
aiofiles==0.6.0
aiohttp==3.6.2
aiosqlite==0.15.0
anyio==2.0.0
asteval==0.9.19
astroid==2.4.2
astunparse==1.6.3
async-timeout==3.0.1
asyncclick==7.1.2.1
asyncpg==0.21.0
attrs==20.2.0
Babel==2.8.0
braceexpand==0.1.5
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
ciso8601==2.1.3
covid-modeling==0.1.1
covid-seird==0.0.7
COVID19Py==0.3.0
cycler==0.10.0
discord.py==1.5.1
future==0.18.2
humanize==2.6.0
idna==2.10
import-expression==1.1.4
iso8601==0.1.13
isort==5.6.4
jishaku==1.19.1.200
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lmfit==1.0.1
matplotlib==3.3.2
mccabe==0.6.1
multidict==4.7.6
numpy==1.19.4
pandas==1.1.4
parsedatetime==2.6
Pillow==8.0.1
pkg-resources==0.0.0
psutil==5.7.3
pybabel==0.0.0.dev0
pycparser==2.20
pydantic==1.6.1
PyJWT==1.7.1
pylint==2.6.0
PyNaCl==1.3.0
pyparsing==2.4.7
PyPika==0.39.1
python-dateutil==2.8.1
python-rapidjson==0.9.1
pytz==2020.4
requests==2.24.0
scipy==1.5.3
selenium==3.141.0
six==1.15.0
sniffio==1.1.0
statcord.py==3.0.5
toml==0.10.1
tortoise-orm==0.16.15
twilio==6.46.0
typing-extensions==3.7.4.3
uncertainties==3.1.4
urllib3==1.25.11
uvloop==0.14.0
wrapt==1.12.1
yarl==1.5.1
folium==0.11.0
h5py==2.10.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
matplotlib==3.2.2
numpy==1.19.0
opencv-python==4.3.0.36
pandas==1.0.5
plotly==4.8.2
tensorflow==2.2.0
tensorflow-estimator==2.2.0
urllib3==1.25.9
